+++ math.c
@@ -3,7 +3,7 @@
 
 int PrintVec3f(Vec3f vec)
 {
-    return printf("{ x: %f, y: %f, z: %f }", vec.x, vec.y, vec.z);
+    return printf("{ x: %f, y: %f, z: %f }", (double)vec.x, (double)vec.y, (double)vec.z);
 }
 
 Vec3f NormalizeVec3f(Vec3f vec)
